backend/requirements.txt
black==22.6.0
Django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
flake8==5.0.4
google-api-core==2.8.2
google-auth==2.10.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.1
google-cloud-storage==2.5.0
google-resumable-media==2.3.3
reorder-python-imports==3.8.2
psycopg2==2.9.3
pyflakes==2.5.0
PyJWT==2.4.0
python-dotenv==0.20.0
requests==2.28.1
urllib3==1.26.11
gunicorn==20.1.0
asgiref==3.5.2
whitenoise
pandas==1.4.3
gspread==5.4.0
django-cors-headers==3.13.0